pyproject.toml
[project]
name = "dom_scrapping_tutorial"
version = "0.1.0"
description = "Dom Scrapping Tutorial"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "faker>=33.1.0",
    "loguru>=0.7.3",
    "mistralai>=1.2.5",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.27.1",
    "streamlit>=1.41.1",
]

[dependency-groups]
test = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]
dev = [
    "ipykernel>=6.29.5",
    "ruff>=0.8.4",
]

[tool.ruff]
# Set the maximum line length to 79.
line-length = 80

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
extend-ignore = ["E501"]
